requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.17
openai>=1.54.0
pillow>=11.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
aiofiles>=24.1.0
python-json-logger>=3.2.0

